rename directories and move definitions
