+++ server/src/lib/cross-origin/multiorigin-cors.ts
@@ -1,4 +1,6 @@
 import type { Request, Response, RequestHandler } from "express";
+import { validate } from "./share";
+import type { Config } from "./share";
 
 /* expected usecase:
 
@@ -11,12 +13,6 @@ app.use(cors.serverSide(corsConfig));
 
 */
 
-type CorsConfig = {
-  origins: string[];
-  methods?: string[];
-  credentials?: boolean;
-};
-
 // express middlewares.
 // ref: https://expressjs.com/ja/guide/using-middleware.html
 
@@ -25,7 +21,7 @@ type CorsConfig = {
 // use the serverSideBlocking() below for CSRF.
 // more about CORS:
 // - https://developer.mozilla.org/ja/docs/Web/HTTP/CORS
-function corsPolicy(config: CorsConfig): RequestHandler {
+function corsPolicy(config: Config): RequestHandler {
   validate(config);
   const {
     origins,
@@ -72,66 +68,5 @@ function corsPolicy(config: CorsConfig): RequestHandler {
   }
 }
 
-function serverSideBlocking(config: CorsConfig) {
-  validate(config);
-  return function(req: Request, res: Response, next: () => void) {
-    if (!req.header("Origin")) {
-      // no origin header == no cors == same origin
-      next();
-      return
-    }
-    const reqOrigin = req.header("Origin");
-    if (!config.origins.some((o) => reqOrigin === o) && config.origins[0] !== "*") {
-      res.status(403).send("unknown origin header: " + reqOrigin);
-      return
-    }
-    next();
-  }
-}
 
-// make the cors config valid.
-// throws error if unrecoverable.
-function validate(config: CorsConfig) {
-  config.credentials = !!config.credentials; // make it boolean. not using Boolean() or new Boolean() because I don't trust JS
-
-  if (config.methods)
-    config.methods = config.methods.map(s => s.toUpperCase());
-
-  // normalize allowOrigin URLs
-  config.origins = config.origins.map(origin => {
-    const url = new URL(origin)
-    if (url.origin === "null") {
-      console.log(`invalid URL: ${origin}. Please prefix this with http:// or https:// if you haven't.`);
-      throw "";
-    }
-    return url.origin
-  });
-
-  // GET and HEAD must be in this field. POST is for convenience.
-  // ref: not found
-  const defaultAllowedMethods = ["GET", "HEAD", "POST"];
-  if (!config.methods) {
-    config.methods = defaultAllowedMethods;
-  }
-  for (const defaultMethod of defaultAllowedMethods) {
-    if (!config.methods.some((m => m === defaultMethod))) {
-      config.methods.push(defaultMethod);
-    }
-  }
-  assertValidConfig(config);
-}
-
-// this throws error if config is not good
-function assertValidConfig(config: CorsConfig) {
-  if (config.origins.length === 0) {
-    throw new Error("Empty allowedOrigins in CORS config: " + JSON.stringify(config));
-  }
-}
-
-export default {
-  serverSide: serverSideBlocking,
-  clientSide: corsPolicy,
-} as {
-  serverSide: (c: CorsConfig) => RequestHandler;
-  clientSide: (c: CorsConfig) => RequestHandler,
-};
+export default corsPolicy;
