+++ server/src/lib/cross-origin/share.ts
@@ -1,0 +1,49 @@
+type Config = {
+  origins: string[];
+  methods?: string[];
+  credentials?: boolean;
+};
+
+// make the cors config valid.
+// throws error if unrecoverable.
+function validate(config: Config) {
+  config.credentials = !!config.credentials; // make it boolean. not using Boolean() or new Boolean() because I don't trust JS
+
+  if (config.methods)
+    config.methods = config.methods.map(s => s.toUpperCase());
+
+  // normalize allowOrigin URLs
+  config.origins = config.origins.map(origin => {
+    const url = new URL(origin)
+    if (url.origin === "null") {
+      console.log(`invalid URL: ${origin}. Please prefix this with http:// or https:// if you haven't.`);
+      throw "";
+    }
+    return url.origin
+  });
+
+  // GET and HEAD must be in this field. POST is for convenience.
+  // ref: not found
+  const defaultAllowedMethods = ["GET", "HEAD", "POST"];
+  if (!config.methods) {
+    config.methods = defaultAllowedMethods;
+  }
+  for (const defaultMethod of defaultAllowedMethods) {
+    if (!config.methods.some((m => m === defaultMethod))) {
+      config.methods.push(defaultMethod);
+    }
+  }
+  assertValidConfig(config);
+}
+
+// this throws error if config is not good
+function assertValidConfig(config: Config) {
+  if (config.origins.length === 0) {
+    throw new Error("Empty allowedOrigins in CORS config: " + JSON.stringify(config));
+  }
+}
+
+export {
+  validate,
+  Config,
+}
